{{cookiecutter.directory_name}}/requirements-test.txt
-r requirements.txt
autopep8
model-bakery
pytest
pytest-cov
pytest-django
requests-mock
ipython